ignore createdAt and lastModified in entity hashes
Mark createdAt and lastModified as auxiliary so they are ignored in `hashCode()`. Both properties do not change the entity, they are just optional metadata.

If the properties are missing, they are set to "now". As a result, the entity hash changes for two data instances even if there is effectively no change in the entity. This leads to unnecessary reloads.